backend/requirements.txt
alembic==1.16.4
APScheduler==3.11.0
async-lru==2.0.5
asyncmy==0.2.10
DateTime==5.5
fastapi[all]==0.116.1
loguru==0.7.3
openpyxl==3.1.5
pandas==2.3.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
Pillow==11.3.0
psutil==7.0.0
pydantic-validation-decorator==0.1.4
PyJWT[crypto]==2.10.1
PyMySQL==1.1.1
redis==6.4.0
SQLAlchemy[asyncio]==2.0.43
sqlglot[rs]==27.8.0
user-agents==2.2.0
